Fix int_ranges
