+++ py-polars/src/functions/eager.rs
@@ -1,10 +1,8 @@
 use polars::functions;
 use polars_core::prelude::*;
-use polars_core::with_match_physical_integer_polars_type;
-use polars_ops::series::new_int_range;
 use pyo3::prelude::*;
 
-use crate::conversion::{get_df, get_series, Wrap};
+use crate::conversion::{get_df, get_series};
 use crate::error::PyPolarsErr;
 use crate::{PyDataFrame, PySeries};
 
@@ -92,30 +90,4 @@ pub fn concat_df_horizontal(dfs: &PyAny) -> PyResult<PyDataFrame> {
 
     let df = functions::concat_df_horizontal(&dfs).map_err(PyPolarsErr::from)?;
     Ok(df.into())
-}
-
-#[pyfunction]
-pub fn eager_int_range(
-    lower: &PyAny,
-    upper: &PyAny,
-    step: &PyAny,
-    dtype: Wrap<DataType>,
-) -> PyResult<PySeries> {
-    let dtype = dtype.0;
-    if !dtype.is_integer() {
-        return Err(PyPolarsErr::from(
-            polars_err!(ComputeError: "non-integer `dtype` passed to `int_range`: {:?}", dtype),
-        )
-        .into());
-    }
-
-    let ret = with_match_physical_integer_polars_type!(dtype, |$T| {
-        let start_v: <$T as PolarsNumericType>::Native = lower.extract()?;
-        let end_v: <$T as PolarsNumericType>::Native = upper.extract()?;
-        let step: i64 = step.extract()?;
-        new_int_range::<$T>(start_v, end_v, step, "literal")
-    });
-
-    let s = ret.map_err(PyPolarsErr::from)?;
-    Ok(s.into())
 }
