+++ py-polars/tests/unit/functions/range/test_int_range.py
@@ -260,3 +260,11 @@ def test_int_range_non_int_dtype() -> None:
         pl.ComputeError, match="non-integer `dtype` passed to `int_range`: String"
     ):
         pl.int_range(0, 3, dtype=pl.String, eager=True)
+
+
+# https://github.com/pola-rs/polars/issues/15307
+def test_int_ranges_non_int_dtype() -> None:
+    with pytest.raises(
+        pl.ComputeError, match="non-integer `dtype` passed to `int_ranges`: String"
+    ):
+        pl.int_ranges(0, 3, dtype=pl.String, eager=True)
